requirements.txt
flask
flask-cors
gtts
google-generativeai
langchain-google-genai 
langchain
numpy 
python-dotenv
